backend/requirements.txt
langchain
openai
groq
python-dotenv
fastapi
uvicorn
mcp-use
motor
